chore: 🔧 Flutter 및 Gradle 버전 업데이트 및 코드 정리
- Flutter SDK 버전 3.35.3에서 3.35.5로 업데이트
- Gradle 버전 8.10.2에서 9.0.0으로 업데이트
- 불필요한 build.yaml 파일 삭제
- 분석 옵션 및 VSCode 설정 파일 수정
- 코드 포맷팅 개선 및 불필요한 코드 삭제

+++ packages/coui_flutter/lib/src/components/control/scrollbar.dart
@@ -151,7 +151,7 @@ class _CoUIScrollbarState extends RawScrollbarState<_CoUIScrollbar> {
         themeValue: compTheme?.radius,
         widgetValue: widget.radius,
       )
-      ..minLength = _kScrollbarMinLength
+      ..minLength = _kScrollbarMinLength.toDouble()
       ..padding = MediaQuery.paddingOf(context) + EdgeInsets.all(_theme.scaling)
       ..scrollbarOrientation = widget.scrollbarOrientation
       ..ignorePointer = !enableGestures;
