+++ packages/coui_flutter/lib/src/components/display/circular_progress_indicator.dart
@@ -245,7 +245,7 @@ class CircularProgressIndicator extends StatelessWidget {
                     backgroundColor: effectiveBackgroundColor,
                     color: effectiveColor,
                     strokeWidth: effectiveStrokeWidth,
-                    value: value,
+                    value: value.toDouble(),
                     valueColor: AlwaysStoppedAnimation<Color>(effectiveColor),
                   ),
                 ),
